add extension name judgement

+++ fileOperator.py
@@ -4,7 +4,6 @@ import shutil
 class fileOperator:
     rootPath = ""
     processedFolder = "no folder with name of Processed or processed"
-    count = 0
     hasValidFolder = 0
     def __init__(self, path):
         self.rootPath = path
@@ -24,14 +23,16 @@ class fileOperator:
             return "Please input valid path\n"
         if self.findProcessedFolder() != 0 :
             return "There is no Processed folder under current path\n"
+        count = 0
         processedFiles = os.listdir(self.processedFolder)
         for f in processedFiles:
             extension = os.path.splitext(f)[1]
             fileName = f.rpartition("_")[0]
             src = self.rootPath + os.sep + fileName + extension
-            self.count = self.count + 1
-            try:
-                shutil.copy(src,self.processedFolder)
-            except IOError:
-                return "Error info: src: " + src + "\n"
-        return str(self.count) + " files are copied to Processed folder\nCleaning suceed\n"
+            if extension == ".JPG" and os.path.exists(src):
+                count = count + 1
+                try:
+                    shutil.copy(src,self.processedFolder)
+                except IOError:
+                    return "Error info: src: " + src + "\n"
+        return str(count) + " files are copied to Processed folder\nCleaning suceed\n"
